create route for update gallery

+++ app/api/gallery_routes.py
@@ -14,13 +14,14 @@ def get_gallery(galleryId):
     gallery = Gallery.query.get(galleryId)
     
     password = request.args.get("p") if request.args.get("p") else ""
-    print("PASSWORD: ", password)
     
     if not gallery:
         return { "errors": "Gallery not found"}, 404
     
+    if current_user.id == gallery.owner_id:
+        return {"gallery": gallery.to_dict_full()}
+    
     if gallery.hashed_password is not None:
-        print("CHECKING_PASSWORD", gallery.hashed_password, gallery.check_password(password))
         if gallery.check_password(password):
             return { "gallery": gallery.to_dict_full()}
         else:
@@ -37,9 +38,11 @@ def post_gallery():
     if form.validate_on_submit():
         gallery = Gallery(
             title = form.data["title"],
-            password = form.data["password"],
             owner_id = current_user.id
         )
+        
+        if form.data["password"]:
+            gallery.password = form.data["password"]
         
         if form.data["date"]:
             gallery.date = date.fromisoformat(str(form.data["date"]))
@@ -52,7 +55,7 @@ def post_gallery():
             if "url" not in upload:
                 return upload, 500
             
-            gallery.preview = f"{CLOUDFRONT_URL}/{preview.filename}"
+            gallery.preview_url = f"{CLOUDFRONT_URL}/{preview.filename}"
         
         db.session.add(gallery)
         db.session.commit()
@@ -76,20 +79,25 @@ def update_gallery(galleryId):
         return { "errors": "Gallery not found"}, 404
     
     if form.validate_on_submit():
-        if form.data["title"]:
-            gallery.title = form.data["title"]
-        if form.data["date"]:
-            gallery.date = form.data["date"]
-        if form.data["password"]:
-            gallery.password = form.data["password"]
-        if form.data["is_public"]:
-            gallery.is_public = form.data["is_public"]
+        gallery.title = form.data["title"]
+        gallery.date = form.data["date"]
+        gallery.password = form.data["password"]
+        if form.data["preview"]:
+            preview = form.data["preview"]
+            preview.filename = get_unique_filename(preview.filename)
+            upload = upload_file_to_s3(preview)
+            
+            if "url" not in upload:
+                return upload, 500
+            
+            gallery.preview_url = f"{CLOUDFRONT_URL}/{preview.filename}"
+            
         
         db.session.commit()
-        return gallery.to_dict()
+        return gallery.to_dict_full()
         
     if form.errors:
-        return {'errors': validation_errors_to_error_messages(form.errors)}, 400
+        return {'errors': form.errors}, 400
     
 @gallery_routes.route("/<int:galleryId>", methods=["DELETE"])
 @login_required
@@ -99,6 +107,15 @@ def delete_gallery(galleryId):
     
     if not gallery or gallery.owner_id != current_user.id:
         return { "errors": "Gallery not found"}, 404
+    
+    delete_preview = remove_file_from_s3(gallery.preview_url)
+    if delete_preview is not True:
+        return delete_preview
+    
+    for item in gallery.items:
+        delete_item = remove_file_from_s3(item.media_url)
+        if delete_item is not True:
+            return delete_item
     
     db.session.delete(gallery)
     db.session.commit()
